resource-sync-worker/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
kubernetes==29.0.0
GitPython==3.1.42
